flask_assignment4/requirements.txt
tensorflow==1.13.1
numpy==1.16.2
Flask==1.0.3
Keras==2.2.4
Pillow==6.2.0
